[FIX] typo and  documentation.
